+++ kite/src/main/kotlin/com/tang/kite/transaction/jdbc/JdbcTransaction.kt
@@ -2,6 +2,7 @@ package com.tang.kite.transaction.jdbc
 
 import com.tang.kite.session.TransactionIsolationLevel
 import com.tang.kite.transaction.Transaction
+import org.slf4j.LoggerFactory
 import java.sql.Connection
 import javax.sql.DataSource
 
@@ -10,13 +11,14 @@ import javax.sql.DataSource
  */
 class JdbcTransaction : Transaction {
 
-    private val connection: Connection
+    private val logger = LoggerFactory.getLogger(JdbcTransaction::class.java)
+
     private lateinit var dataSource: DataSource
+    private lateinit var connection: Connection
     private var isolationLevel: TransactionIsolationLevel? = null
     private var autoCommit: Boolean = false
 
     constructor(dataSource: DataSource, isolationLevel: TransactionIsolationLevel?, autoCommit: Boolean) {
-        this.connection = dataSource.connection
         this.dataSource = dataSource
         this.isolationLevel = isolationLevel
         this.autoCommit = autoCommit
@@ -27,22 +29,41 @@ class JdbcTransaction : Transaction {
     }
 
     override fun getConnection(): Connection {
-        isolationLevel?.let {
-            connection.transactionIsolation = it.level
+        if (::connection.isInitialized.not() || connection.isClosed) {
+            openConnection()
         }
-        this.connection.autoCommit = autoCommit
         return connection
     }
 
+    private fun openConnection() {
+        val start = System.currentTimeMillis()
+        connection = dataSource.connection
+        if (logger.isDebugEnabled) {
+            logger.debug("Opening JDBC Connection [{}] in {} ms", connection, System.currentTimeMillis() - start)
+        }
+        isolationLevel?.let {
+            connection.transactionIsolation = it.level
+        }
+        logger.debug("Setting autocommit to [{}] on JDBC Connection [{}]", autoCommit, connection)
+        this.connection.autoCommit = autoCommit
+    }
+
     override fun commit() {
+        logger.debug("Committing JDBC Connection [{}]", connection)
         connection.commit()
     }
 
     override fun rollback() {
+        logger.debug("Rolling back JDBC Connection [{}]", connection)
         connection.rollback()
     }
 
     override fun close() {
+        if (connection.autoCommit.not()) {
+            logger.debug("Resetting autocommit to true on JDBC Connection [{}]", connection)
+            connection.autoCommit = true
+        }
+        logger.debug("Closing JDBC Connection [{}]", connection)
         connection.close()
     }